Add support for psa_trusted_storage_api v1.0.0 APIs (#2)
* Adopt v1.0.0 of the psa_trusted_storage_api defined protected_storage.h API

The following provides more information on this commit.
- Adopt the v1.0.0 release version of the protected_storage.h API.
- Fix psa-storage-example-app to adopt v1.0.0 protected_storage.h.
- The following points provide useful information on the origin of the
  API header files included in this commit. Related software components
  (e.g. mbed-crypto, mbed-os, psa-arch-test) have adopted pre-release versions
  which will move to the v1.0.0 in future releases.
- Marcus Streets (the gatekeeper for the psa_trusted_storage_api specification
  repo) has applied the v1.0.0 tag to psa_trusted_storge_api repo. This is at
  the following commit:
    SHA: c9853c2105054256647faf35df7ad09be8296571
  WARNING: do not use the v1.0 tag on the repo.
- protected_storage.h includes error.h which is defined in:
    https://github.com/ARMmbed/PSA-IPC-doc/blob/master/include/psa/error.h
  Hence a psa_trusted_storage_api release depends on a PSA-IPC-doc release.
  The v1.0.0 tag has also been applied on PSA-IPC-doc repo, at the following
  commit:
    SHA: ff400a05d5f6385cac64c521981cd18bb5a57d7a
  Note that for this release the same tag (v1.0.0) has been used on the
  matching release versions of psa_trusted_storage_api and PSA-IPC-doc.
- protected_storage.h includes storage_common.h. The v1.0.0 version of
  storage_common.h is also included is also included in this commit.
- Note the *.h files mentioned here may have trailing white spaces. This has
  not been removed (the files have not been modified in any way from
  the release v1.0.0 versions).

* Add mbed-crypto config.h called mbed_crypto_psa_storage_config.h.

The following provides more information about this commit:
- Add mbed_crypto_psa_storage_config.h. This is a copy of the following file:
    mbed-crypto/include/mbedtls/config.h
  This file will be used to configure a build of mbed-crypto to
  use the psa-storage backend.
- Add check_config.h. This is a copy of the following file:
    mbed-crypto/include/mbedtls/check_config.h
- These files were taken from the mbed-crypto project at the following SHA:
    SHA: 47f2de132936905d97a93e2ddf7f5237ab232fbe
  from v1.1.0d2.

* Implement release v1.0.0 internal_trusted_storage.h storage backend.

The following provides more information on this commit.
- Add the release v1.0.0 version of internal_trusted_storage.h.
- Add code to implement the API by wrapping psa_ps_xxx() calls
  to the protected_storage API.
- Remove psa_storage_type.h as this is no longer used.
- Update mbed_crypto_psa_storage_config.h to not define
  MBEDTLS_PSA_ITS_FILE_C, and define new symbols. This config file configures
  the mbed-crypto build to use this library as a persistent storage backend.
  See later for more details.

In order for mbed-crypto to use the psa-storage library as a backend, the
mbed-crypto build must be configured correctly. The following instructions
are provided for this purupose:
- Provide Makefile flags so mbed-crypto links with the psa-storage library
  e.g.:
    export TOPDIR=${PWD}
    make -C psa-storage
    make -C psa-storage install prefix=$(TOPDIR)/psa-storage/usr/local
    export CFLAGS="-I${TOPDIR}/psa-storage/inc \
                   -I${TOPDIR}/psa-storage/configs \
                   -DMBEDTLS_CONFIG_FILE='<psa_storage_config.h>'"
    export LDFLAGS="-L${TOPDIR}/psa-storage/usr/local/lib -lpsastorage"
    make -C mbed-crypto CFLAGS=${CFLAGS} LDFLAGS=${LDFLAGS}
- The 1st step defined TOPDIR.
- The 2nd step builds the psa-storage project.
- The 3rd step installs the psa-storage build artifacts. These will be used
  when building mbed-crypto.
- The 4th step defines CFLAGS to be used later.
- The 5th step defines LDFLAGS to be used later.
- The 6th step builds mbed-crypto with the defined CFLAGS and LDFLAGS.
- The mbed_crypto_psa_storage_config.h file is the config.h used when building
  mbed-crypto. Note the MBEDTLS_PSA_ITS_FILE_C symbol is not set in this file.
  The mbed-crypto psa_its_file.c module provides an internal persistent file
  storage implementation using a pre-v1.0.0 version of
  internal_trusted_storage.h at its upper edge. This code is conditionally
  compiled into a build if the config.h MBEDTLS_PSA_ITS_FILE_C symbol is set.
  In order to use the psa-storage backend this symbol should therefore not be set.
- The mbed_crypto_psa_storage_config.h file includes the
  MBEDTLS_PSA_ITS_FILE_C_INCLUDE_TEST symbol definiton. This results in useful
  mbed-crypto storage related tests being built even when
  MBEDTLS_PSA_ITS_FILE_C is not set.
- These are the useful mbed-crypto tests from a storage perspective:
  - test_suite_psa_its. This tests the psa_crypto_its.h interface
    which is equivalent of the PSA internal_trusted_storage.h API.
    Hences its a bottom-half test.
  - test_suite_persistent_key. This tests the storage of keys and
    hences its a top-half test.

* PSA Storage: Add error.h Apache 2.0 License header as agreed by Adrian Shaw.
  Adrian Shaw is the gatekeeper for the PSA-IPC-doc repo from which a version of the
  error.h was taken.

+++ app/main.c
@@ -467,7 +467,6 @@ int hexcmp( uint8_t * a, uint8_t * b, uint32_t a_len, uint32_t b_len )
 
 #define TEST_SUITE_ACTIVE
 
-#include "psa/psa_storage_types.h"
 #include "psa/protected_storage.h"
 
 /* Internal definitions of the implementation, copied for the sake of
@@ -521,14 +520,14 @@ static psa_status_t psa_ps_set_wrap( psa_storage_uid_t uid,
 {
     if( uid_max != (psa_storage_uid_t)( -1 ) && uid_max < uid )
         uid_max = uid;
-    return( psa_ps_set( uid, data_length, p_data, create_flags ) );
+    return( psa_ps_set( uid, (size_t) data_length, p_data, create_flags ) );
 }
 
 void test_set_get_remove( int uid_arg, int flags_arg, data_t *data )
 {
     psa_storage_uid_t uid = uid_arg;
     uint32_t flags = flags_arg;
-    uint32_t ret_len = 0;
+    size_t ret_len = 0;
     struct psa_storage_info_t info;
     unsigned char *buffer = NULL;
 
@@ -562,7 +561,7 @@ void test_set_overwrite( int uid_arg,
     psa_storage_uid_t uid = uid_arg;
     uint32_t flags1 = flags1_arg;
     uint32_t flags2 = flags2_arg;
-    uint32_t ret_len = 0;
+    size_t ret_len = 0;
     struct psa_storage_info_t info;
     unsigned char *buffer = NULL;
 
@@ -603,7 +602,7 @@ void test_set_multiple( int first_id, int count )
     psa_storage_uid_t uid;
     char stored[40];
     char retrieved[40];
-    uint32_t ret_len = 0;
+    size_t ret_len = 0;
 
     memset( stored, '.', sizeof( stored ) );
     for( uid = uid0; uid < uid0 + count; uid++ )
@@ -670,7 +669,7 @@ void test_get_at( int uid_arg, data_t *data,
     size_t length = length_arg >= 0 ? length_arg : 0;
     unsigned char *trailer;
     size_t i;
-    uint32_t ret_len = 0;
+    size_t ret_len = 0;
 
     ASSERT_ALLOC( buffer, length + 16 );
     trailer = buffer + length;
